完善进度输出内容

+++ common/src/main/java/cn/org/hentai/squeeze/common/util/BytesUnit.java
@@ -5,13 +5,13 @@ package cn.org.hentai.squeeze.common.util;
  */
 public final class BytesUnit
 {
-    public static String convert(int BPS)
+    public static String convert(long BPS)
     {
         char unit = 'k';
-        float speed = 0;
-        if (BPS > 1024 * 1024 * 1024) { speed = BPS / 1024f / 1024 / 1024; unit = 'G'; }
-        else if (BPS > 1024 * 1024) { speed = BPS / 1024f / 1024; unit = 'M'; }
-        else speed = BPS / 1024f;
+        double speed = 0d;
+        if (BPS > 1024 * 1024 * 1024) { speed = BPS / 1024d / 1024 / 1024; unit = 'G'; }
+        else if (BPS > 1024 * 1024) { speed = BPS / 1024d / 1024; unit = 'M'; }
+        else speed = BPS / 1024d;
         return String.format("%.2f%s", speed, unit);
     }
 }
